+++ react/components/ProductAssemblyOptions/ProductAssemblyOptionItemCustomize.tsx
@@ -97,6 +97,19 @@ interface ButtonProps {
   collapse?: 'left' | 'right' | 'none'
 }
 
+const hasValues = (itemChildren: any) => {
+  const [name] = Object.getOwnPropertyNames(itemChildren)
+  const inputs = itemChildren[name].inputValues
+
+  if (!itemChildren[name]?.valuesOfInputValues) {
+    return false
+  }
+
+  return inputs.findIndex((input: { label: string }) => {
+    return itemChildren[name].valuesOfInputValues[input.label] !== ''
+  }) !== -1
+}
+
 const ProductAssemblyOptionItemCustomize: FC<Props> = ({
   children,
   buttonProps = {},
@@ -117,7 +130,10 @@ const ProductAssemblyOptionItemCustomize: FC<Props> = ({
     return null
   }
 
-  const handleClick = () => {
+  const handleClick = (e: any) => {
+    if(hasValues(itemChildren)) {
+      e.stopPropagation()
+    }
     setModalOpen(true)
   }