+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Stop click propagation when input has values to avoid unchecking accidentally
+
 ## [2.11.3] - 2022-07-19
 
 ### Fixed